pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "patchwise"
version = "0.2.1"
description = "A tool for reviewing kernel patches"
authors = [{ name = "David Gantman", email = "dgantman@qti.qualcomm.com" }]
requires-python = ">=3.10"
dependencies = [
    "Sphinx>=8.1.3",
    "litellm>=1.74",
    "cachetools>=5.5.2",
    # "dtschema>=2025.2",
    "yamllint>=1.26.3",
    "ply>=3.11",
    "GitPython>=3.1.44",
    "codespell>=2.4.1",
    "pytest>=8.3.5",
    "rich-argparse>=1.7.0",
    "tqdm>=4.67.1",
    "platformdir"
]

[project.scripts]
patchwise = "patchwise.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["patchwise*"]

[tool.setuptools.package-data]
patchwise = ["patches/**"]
